Add files via upload
Yoda is building the schedule of the curriculum for his apprentices by assigning periods to courses in a way that the load during each period is reasonably balanced. Yoda would like to satisfy the following constraints:

- The Jedi curriculum is defined by a set of courses and a set of prerequisite relationships among them.

- There is a number of available periods.

- Each course has assigned a number of units.

- An apprentice must take a minimum number of units and cannot exceed a maximum number of units during each period.

An apprentice must take a minimum number of courses and cannot exceed a maximum number of courses during each period.

Yoda has been taken by an important mission to rescue princess Leia and has asked you to assign a period to every course in a way that the minimum and maximum units for each period, the minimum and maximum number of courses for each period, and the prerequisite relationships are satisfied. Moreover, you must minimize the maximum number of courses taken during each period. Yoda told you that you can only use arithmetic and reified constraints on this assignment.