Add files via upload
Mastermind is a game where we try to identify the secret code which is a combination of 4 colors. The colors range between numbers 1-6 representing 6 different colors. We are awarded a "Correct" point if the number in the code guess is at the same place in the actual secret code, as well as an "Incorrect" point if the number exists somewhere in the code but not in the same place as the guessed code. Also, this "incorrect" award is given only once if the number is repeated in the code. 
The above code solves this problem and deduces the secret code under 5 minutes.