Add files via upload
Han Solo is heading a major relief operation across the galaxy. He needs to deliver relief goods to
many planets that were attacked by the imperial fleet. The deliveries are achieved by dispatching
a fleet of X-wing fighters from the ice planet Hoth. He must design a route for each fighter so that
all of the planets are served by exactly one fighter. The fighter have a fixed storage capacity and
the planets have different demands in relief goods. Figures 1 and 2 illustrate a small instance and
a feasible solution to that problem. The planets are labeled from 0 to 4, with 0 being Hoth. The
solution uses two fighters which are indicated by different colored routes.
Solo must minimize a lexicographic objective function whose first component minimizes the time at
which the latest planet gets its relief and the second component minimizes the sums of the trip times
of all fighters.