Add files via upload
Han Solo has been appointed by Obi-Wan to design a resilient communication system for the
Republic. He needs to assign a number of control centers to different locations under a variety of
constraints, so that they can communicate efficiently. Han Solo has received the following data.

1. C: the set of control centers;
2. L: the set of locations;
3. f : a frequency matrix that describes how much the control centers communicate with each
other. In other words, fi,j describes the communication frequency between control centers i
and j;
4. h: a hop matrix over the locations that describes the distance in hops between two locations.
In other words, hi,j describes the hop distance between locations n and m;
5. Separed: a collection of control center sets S1, ..., Sk. Each set Si specifies that its control
centers cannot be located at the same location.
6. Together: a collection of control center sets S1, ..., Sk. Each set Si specifies that its control

centers must be located at the same location.
The goal is to assign the control centers to locations so that the separation and togetherness
constraints are satisfied and each location is assigned at most 5 control centers. In addition, the
objective is to minimize the cost of communication between the control centers. The communication
cost between two control centers i and j is the product of the frequency fi,j by the number of hops
between the locations of control centers i and j.