Add files via upload
Hospitals are getting organized and are careful to isolate
COVID19 patients from others. General Leia is in charge of allocating residents to patients. She
has organized the hospital in zones to separate patients. Each patient is allocated to a zone and
has a load that describes how sick she/he is. General Leia wants to assign residents to patients so
that the following constraints are satisfied:
1. Each resident is assigned to at least minPatients and at most maxPatients;
2. Each resident cannot have a load that exceeds maxLoad, where the load of a resident is the
sum of the loads of her patients;
3. Each resident only works in one zone.
General Leia also wants to make sure that the load of the residents is as balanced as possible, i.e.,
that the maximum difference in load between any two residents is minimized.
Help General Leia solve this problem!