Add files via upload
The resistance is building houses on Takodana. Each house has a number
of tasks linked by precedence constraints and a release date dictated
by when the material arrives. There are earliness and tardiness costs
associated with some tasks of each house. Each task has a fixed
duration. The goal is to minimize the sum of the tardiness costs,
while making sure that the earliness costs do not exceed the value
maxEarlinessCost. Can you help the resistance?